+++ lib/src/dartarabic.dart
@@ -1,6 +1,7 @@
 import 'operations.dart';
 
-abstract class DartArabic {
+class DartArabic {
+  const DartArabic._();
   /// Strip Harakat from [text] arabic word except Shadda. The striped marks are : FATHA, DAMMA, KASRA,  FATHATAN, DAMMATAN, KASRATAN,SUKUN Marks,
   static String stripHarakat(String text) => ArOp.strip_harakat(text);