+++ lib/src/arabic/arabic_symbols.dart
@@ -2,7 +2,48 @@ import 'package:dartarabic/src/arabic/arabic.dart';
 import 'package:dartarabic/src/unicode_data/script.dart';
 
 class ArabicSymbols {
-  ArabicSymbols();
+  ArabicSymbols() {
+    UTHMANI_STOP_ZIM = UTHMANI_STOP_JEEM;
+    SUKUN_MARKS = [SUKUN, SUKUN_CURVY, SUKUN_ROUNDED];
+    HARAKAT = [
+      FATHATAN,
+      DAMMATAN,
+      KASRATAN,
+      FATHA,
+      DAMMA,
+      KASRA,
+      ...SUKUN_MARKS,
+    ];
+    HARAKAT_PATTERN = RegExp(r'[' + HARAKAT.join() + r']', unicode: true);
+    SHORTHARAKAT = [FATHA, DAMMA, KASRA, SUKUN];
+    ALL_UTHMANI_STOPS = [
+      UTHMANI_STOP_SAAD_LAM,
+      UTHMANI_STOP_QAF_LAM,
+      UTHMANI_STOP_MEEM,
+      UTHMANI_STOP_LAM,
+      UTHMANI_STOP_JEEM,
+      UTHMANI_STOP_DOTS_THREE,
+      UTHMANI_STOP_SEEN_HIGH,
+    ];
+    TASHKEEL = [
+      FATHATAN,
+      DAMMATAN,
+      KASRATAN,
+      FATHA,
+      DAMMA,
+      KASRA,
+      SUKUN,
+      SHADDA
+    ];
+    TANWIN = [FATHATAN, DAMMATAN, KASRATAN];
+    DIACRITICS = DIACRITICS_RANGE.where((element) {
+      final found = UnicodeScript.scripts
+          .where((script) => element >= script.start && element <= script.end);
+      if (found.isEmpty) return false;
+      final script = found.single;
+      return script.category == "Mn";
+    }).map((e) => String.fromCharCode(e));
+  }
 
   // numbers
   final ZERO = '\u0660';
@@ -24,6 +65,7 @@ class ArabicSymbols {
   final EIGHT = '\u0668';
 
   final NINE = '\u0669';
+
   // ./numbers
 
   final PERCENT = '\u066a';
@@ -45,8 +87,10 @@ class ArabicSymbols {
 
   ///  Quranic Annotation small Tah &#1558;
   final QURAN_ANNOTATION_TAH = '\u0615';
+
   ///  Quranic Annotation small Zain &#1559;
   final QURAN_ANNOTATION_ZAIN = '\u0617';
+
   /// Punctuation Three Dots &#1566;
   final TRIPPLE_DOT = '\u061E';
 
@@ -69,16 +113,17 @@ class ArabicSymbols {
 
   /// QAF-LAM-ALEF-MAKSURA &#1751;
   final UTHMANI_STOP_QAF_LAM = '\u06d7';
+
   /// SMALL HIGH MEEM &#1752;
   final UTHMANI_STOP_MEEM = '\u06d8';
+
   /// HIGH LAM-ALEF U+06D9
   final UTHMANI_STOP_LAM = '\u06d9';
 
   /// SMALL HIGH JEEM &#1754;
   final UTHMANI_STOP_JEEM = '\u06da';
 
-  /// SMALL HIGH JEEM &#1754; [UTHMANI_STOP_JEEM]
-  late final UTHMANI_STOP_ZIM = UTHMANI_STOP_JEEM;
+
 
   /// SMALL HIGH THREE DOTS &#1755;
   final UTHMANI_STOP_DOTS_THREE = '\u06db';
@@ -87,24 +132,16 @@ class ArabicSymbols {
   final UTHMANI_STOP_SEEN_HIGH = '\u06dc';
 
 
-  /// LIST OF ALL UTHMANI STOPS
-  late final List<String> ALL_UTHMANI_STOPS = [
-    UTHMANI_STOP_SAAD_LAM,
-    UTHMANI_STOP_QAF_LAM,
-    UTHMANI_STOP_MEEM,
-    UTHMANI_STOP_LAM,
-    UTHMANI_STOP_JEEM,
-    UTHMANI_STOP_DOTS_THREE,
-    UTHMANI_STOP_SEEN_HIGH,
-  ];
 
   /// AYAH END SIGN &#1757;
   final SIGN_AYAH_END = '\u06DD';
+
   /// STAR OF RUB-EL-HIZEB &#1758;
   final STAR_RUBEL_HIZEB = '\u06DE';
 
   /// SAJDA SIGN &#1769;
   final SIGN_SAJDAH = '\u06E9';
+
 //   Diacritics                      ;
   final FATHATAN = '\u064b';
 
@@ -124,40 +161,24 @@ class ArabicSymbols {
   final SUKUN_CURVY = '\u06e1';
   final SUKUN_ROUNDED = '\u06df';
 
-  late final SUKUN_MARKS = [SUKUN, SUKUN_CURVY, SUKUN_ROUNDED];
-  late final HARAKAT = [
-    FATHATAN,
-    DAMMATAN,
-    KASRATAN,
-    FATHA,
-    DAMMA,
-    KASRA,
-    ...SUKUN_MARKS,
-  ];
-  late final HARAKAT_PATTERN =
-      RegExp(r'[' + HARAKAT.join() + r']', unicode: true);
+  /// SMALL HIGH JEEM &#1754; [UTHMANI_STOP_JEEM]
+  late final UTHMANI_STOP_ZIM;
 
-  late final SHORTHARAKAT = [FATHA, DAMMA, KASRA, SUKUN];
+  late final SUKUN_MARKS;
 
-  late final TASHKEEL = [
-    FATHATAN,
-    DAMMATAN,
-    KASRATAN,
-    FATHA,
-    DAMMA,
-    KASRA,
-    SUKUN,
-    SHADDA
-  ];
+  late final HARAKAT;
 
-  late final TANWIN = [FATHATAN, DAMMATAN, KASRATAN];
+  late final HARAKAT_PATTERN;
+
+  late final SHORTHARAKAT;
+
+  late final TASHKEEL;
+
+  late final TANWIN;
+
+  /// LIST OF ALL UTHMANI STOPS
+  late final List<String> ALL_UTHMANI_STOPS;
 
   final List<int> DIACRITICS_RANGE = [for (int i = 0x0600; i < 0x06ff; ++i) i];
-  late final DIACRITICS = DIACRITICS_RANGE.where((element) {
-    final found = UnicodeScript.scripts
-        .where((script) => element >= script.start && element <= script.end);
-    if (found.isEmpty) return false;
-    final script = found.single;
-    return script.category == "Mn";
-  }).map((e) => String.fromCharCode(e));
+  late final Iterable<String> DIACRITICS;
 }