+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- test case for `stripDiacritics` added
 - removed `unicode_data` dependency. migrated to full null-safety.
 ## [0.0.1-nullsafety.0]
 - migrated to null-safety